introduce `tcc_mark` decorator for connector methods
This introduces a generic way to "mark" a method in the base connector
class. It can then be used, for instance in `__init_subclass__` to
decide what to do with these markers, e.g:

```python
if callable(attrvalue) and is_tcc_marked('require-permission'):
    setattr(cls, attrname, cls.apply_permissions(attrvalue))

```

The only use case we know of, for now, is to tag a base method for later
decoration. Therefore, we might want later to simply decorate the
methods with:

```
@wrap_subclass_implem(apply_permissions)
@wrap_subclass_implem(retry)
def explain(...):
    pass
```

where `retry` and `apply_permissions` would be decorator-like functions.