make `database` parameter mandatory
The fact the `database` was not mandatory was odd to me. As a matter
of fact it works for MSSQL as `pymssql` has a default behaviour.
But it's not the case for `psycopg2`.

For example

```
psycopg2.connect(host='127.0.0.1', port=host_port, user='ubuntu', password='ilovetoucan')
```

would fail without a database !

I chose to put a default value `database = 'postgres'`, which is the default postgres
database. Thanks to this, we should be able to retrieve informations like available
databases !