Feature :: get form (#96)
* make test work on MacOs and unix

* allow to have empty query (default on {})

* move `database` field from MongoConnector to MongoDataSource

* move `db` field from AzureMSSQLConnector to `database` in AzureMSSQLDataSource

* move `db` field from GoogleCloudMySQLConnector to `database` in GoogleCloudMySQLDataSource

* move `database` field from HiveConnector to `database` in HiveDataSource

* move `db` field from MSSQLConnector to `database` in MSSQLDataSource

* move `db` field from MySQLConnector to `database` in MySQLDataSource

* move `db` field from PostgresConnector to `database` in PostgresDataSource

* make `database` parameter mandatory

The fact the `database` was not mandatory was odd to me. As a matter
of fact it works for MSSQL as `pymssql` has a default behaviour.
But it's not the case for `psycopg2`.

For example

```
psycopg2.connect(host='127.0.0.1', port=host_port, user='ubuntu', password='ilovetoucan')
```

would fail without a database !

I chose to put a default value `database = 'postgres'`, which is the default postgres
database. Thanks to this, we should be able to retrieve informations like available
databases !

* Add a `get_form` method on the datasource

This method allows us to retrieve a better form schema, depending
on the current config of a query

* s/contraints/constraints/g

* small flake8 error

* remove connection_params property in favor of get_connection_params(**kwargs)

* Deprecation warning

* dedicated method to create string enums

* add get_form for mysql

* Bump version to 0.23.0

* change get_df_and_count in get_slice

* get_slice :: return a namedtuple

* s/SliceResult/DataSlice/g