MongoConnector :: keep the client alive and cache validate methods
For a bulk query, a very long time was spent opening and closing
again and again the same connection. To avoid this, we now
can open the connection only once and use it for many
requests by using the connector as a context manager.
Furthermore the validations done on database or collection existence were
also taking a lot of time so some cache has been added to avoid
doing the same check many times.