snowflake: Add a default warehouse field
This commit adds a default warehouse field in the Snowflake connector.
A default warehouse will be applied to any datasource that does not have
its warehouse field defined.

This commit only provides a default value for the generated form that
will be displayed on the smallapp, the value has to be submitted from
the form in the browser.