snowflake: fix how query parameters are applied
This commit fixes the query execution in the Snowflake connector; as
reported in the Issue #274
(https://github.com/ToucanToco/toucan-connectors/issues/274), query
parameters were not applied to a query through Snowflake's own client
but through one of our custom functions.

This commit also introduces the use of `fetch_pandas_all` to retrieve
the results of a query in order to improve efficiency on `SELECT`
queries.