Make snowflake connector be able to hold external tokens and credentials
This commit adds three fields to the Snowflake connector :
	* _sso_credentials_access
	* sso_credentials_keeper
	* user_tokens_keeper

These fields will be used in laputa to inject at runtime tokens and
credentials keeper to allow the connector to use these credentials in a
secure way instead of relying on plain text data.

This fields will also allow the connector to update these credentials if
needed.