fix: Ensure STL dates are returned with the right dtype
BigQuery and Athena translators store objects in the SQL DATE type as datetime.date objects.
That is correct, but pandas does not consider datetime.date objects as datetimes. As a result, when
serializing and deserializing a dataframe containing datetime.date objects, these get dumped/loaded
as strings.

This change ensures the right type is used.

TCTC-3424

Signed-off-by: Luka Peschke <luka.peschke@toucantoco.com>