fix(mysql): Do not specify the database for schema-related functions
Both queries (SHOW DATABASES and the more complex one in build_database_model_extraction_query),
do not require a database to function.

Currently, the database was forced to "mysql", which caused the connection to fail in case the database
did not exist or if the SQL user used by the connector did not have sufficient permissions.

TCTC-3768

Signed-off-by: Luka Peschke <luka.peschke@toucantoco.com>